EdgeKey to its own file, and changed distance calculation to not project on XY plane
